refactor: restructure components
