+++ src/components/quiz/create/question/CreateQuestion.tsx
@@ -1,10 +1,11 @@
 import { Button } from '@mui/material';
 import Formsy from 'formsy-react';
-import React, { useState } from 'react';
-import { QuestionType, questionTypes } from '../../../services/quiz/models';
-import DropDown from '../../common/DropDown';
-import TextInput from '../../common/TextInput';
-
+import { useState } from 'react';
+import { QuestionType, questionTypes } from '../../../../services/quiz/models';
+import DropDown from '../../../common/DropDown';
+import TextInput from '../../../common/TextInput';
+import { TypeQuestionCreate } from './TypeQuestionCreate';
+import './CreateQuestion.scss';
 type CreateQuestion = {
   handleClose: () => void;
   createQuestion: (model: QuestionType) => void;
@@ -12,6 +13,8 @@ type CreateQuestion = {
 
 const CreateQuestion = ({ handleClose, createQuestion }: CreateQuestion) => {
   const [createQuestionEnabled, setCreateQuestionEnabled] = useState(false);
+  const [selectedType, setSelectedType] = useState<string>('');
+
   const createQuestionClick = (model: QuestionType) => {
     createQuestion(model);
     handleClose();
@@ -34,17 +37,27 @@ const CreateQuestion = ({ handleClose, createQuestion }: CreateQuestion) => {
           placeholder="Add your question title"
         />
       </div>
-      <DropDown
-        required
-        id="question-type"
-        name="type"
-        label="type"
-        placeholder="Select your type question"
-        items={Object.keys(questionTypes).map((k) => ({
-          value: k,
-          label: questionTypes[k]
-        }))}
-      />
+      <div className="question-type">
+        <div className="type-selector">
+          <DropDown
+            required
+            id="question-type"
+            name="type"
+            label="type"
+            placeholder="Select your type question"
+            onSelect={(value: string | number) =>
+              setSelectedType(value.toString())
+            }
+            items={Object.keys(questionTypes).map((k) => ({
+              value: k,
+              label: questionTypes[k]
+            }))}
+          />
+        </div>
+        <div className="type-create">
+          {selectedType && <TypeQuestionCreate typeQuestion={selectedType} />}
+        </div>
+      </div>
 
       {!createQuestionEnabled && (
         <p className="error">Required fields must be filled</p>
